backend/requirements.txt
# AI Agent
langgraph>=0.5.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
openai>=1.0.0
mem0ai>=0.1.0
tavily-python>=0.5.0

# API
fastapi>=0.115.0
uvicorn>=0.30.0
sse-starlette>=2.0.0
python-multipart>=0.0.9

# Database
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Auth
bcrypt>=4.0.0
PyJWT>=2.0.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
